Added GetFavouriteIdUseCase and SetUserFavouritesUseCase
